Rename util to util.server
