+++ src/analyzer/analyze.server.ts
@@ -9,7 +9,7 @@ import {
   TruckUserConfig,
 } from "./model"
 import { log, setLogLevel } from "./log.server"
-import { describeAsyncJob, formatMs, writeRepoToFile, getRepoName } from "./util"
+import { describeAsyncJob, formatMs, writeRepoToFile, getDirName } from "./util.server"
 import { GitCaller } from "./git-caller"
 import { emptyGitCommitHash } from "./constants"
 import { resolve, isAbsolute, join, sep } from "path"
